# RFID tracking web MVC/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RFID tracking API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/RegularUsers": {
      "get": {
        "tags": [
          "RegularUsers"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegularUsers"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegularUsers"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegularUsers"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RegularUsers"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularUsers"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularUsers"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegularUsers"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegularUsers"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularUsers"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularUsers"
                }
              }
            }
          }
        }
      }
    },
    "/api/RegularUsers/{id}": {
      "get": {
        "tags": [
          "RegularUsers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegularUsers"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularUsers"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularUsers"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RegularUsers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularUsers"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularUsers"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegularUsers"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "RegularUsers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Shooters": {
      "get": {
        "tags": [
          "Shooters"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shooter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shooter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shooter"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shooters"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shooter"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Shooter"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Shooter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Shooter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shooter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shooter"
                }
              }
            }
          }
        }
      }
    },
    "/api/Shooters/{id}": {
      "get": {
        "tags": [
          "Shooters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Shooter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shooter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shooter"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Shooters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shooter"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Shooter"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Shooter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Shooters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Staff": {
      "get": {
        "tags": [
          "Staff"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Staff"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Staff"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Staff"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Staff"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Staff"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Staff"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Staff"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          }
        }
      }
    },
    "/api/Staff/{id}": {
      "get": {
        "tags": [
          "Staff"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Staff"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Staff"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Staff"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Staff"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Staff"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegularUsers": {
        "required": [
          "userId",
          "weaponId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "weaponId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "Shooter": {
        "required": [
          "isPictureIdShown",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "mailAddress": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "format": "tel",
            "nullable": true
          },
          "shooterId": {
            "type": "integer",
            "format": "int32"
          },
          "birthday": {
            "type": "string",
            "format": "date-time"
          },
          "isPictureIdShown": {
            "type": "boolean"
          },
          "directorAcceptedPictureId": {
            "type": "string",
            "format": "uuid"
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Staff": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "mailAddress": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "format": "tel",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}